kyc-backend/requirements.txt
fastapi
uvicorn
python-multipart
fireworks-ai
pydantic
httpx